third_party/WebKit/LayoutTests/bluetooth/requestDevice/consumes-user-gesture.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/testdriver.js"></script>
<script src="../../resources/testdriver-vendor.js"></script>
<script src="../../resources/bluetooth/bluetooth-helpers.js"></script>
<script src="../../external/wpt/bluetooth/resources/bluetooth-helpers.js"></script>
<script>
'use strict';
bluetooth_test(t => {
  return setUpHealthThermometerAndHeartRateDevices()
    .then(() => callWithTrustedClick(() => {
      var first = navigator.bluetooth.requestDevice({
        filters: [{services: ['heart_rate']}]});
      var second = navigator.bluetooth.requestDevice({
        filters: [{services: ['heart_rate']}]});
      return Promise.all([
        first.then(device => assert_equals(
          device.constructor.name, 'BluetoothDevice')),
        promise_rejects(t, 'SecurityError', second)
      ]);
    }));
}, 'Consumes a user gesture.');
</script>
